Refactor accounts into sub package
account/setup.py
account/service.py
account/storage.py
