+++ azurectl/commands/compute_endpoint.py
@@ -60,7 +60,7 @@ options:
 """
 # project
 from base import CliTask
-from ..azure_account import AzureAccount
+from ..account.service import AzureAccount
 from ..utils.collector import DataCollector
 from ..utils.output import DataOutput
 from ..endpoint import Endpoint
