+++ azurectl/commands/compute_request.py
@@ -31,7 +31,7 @@ options:
 """
 # project
 from base import CliTask
-from ..azure_account import AzureAccount
+from ..account.service import AzureAccount
 from ..utils.collector import DataCollector
 from ..utils.output import DataOutput
 from ..request_result import RequestResult
