+++ azurectl/commands/compute_storage.py
@@ -93,7 +93,7 @@ from pytz import utc
 
 # project
 from base import CliTask
-from ..azure_account import AzureAccount
+from ..account.service import AzureAccount
 from ..utils.collector import DataCollector
 from ..utils.output import DataOutput
 from ..logger import log
