+++ azurectl/commands/compute_vm.py
@@ -82,7 +82,7 @@ options:
 """
 # project
 from base import CliTask
-from ..azure_account import AzureAccount
+from ..account.service import AzureAccount
 from ..utils.collector import DataCollector
 from ..utils.output import DataOutput
 from ..logger import log
