+++ test/unit/data_disk_test.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from mock import patch
 from test_helper import *
 
-from azurectl.azure_account import AzureAccount
+from azurectl.account.service import AzureAccount
 from azurectl.azurectl_exceptions import *
 from azurectl.config import Config
 from azurectl.data_disk import DataDisk
