+++ test/unit/endpoint_test.py
@@ -16,7 +16,7 @@ from azure.servicemanagement.models import (
 
 )
 # project
-from azurectl.azure_account import AzureAccount
+from azurectl.account.service import AzureAccount
 from azurectl.azurectl_exceptions import *
 from azurectl.config import Config
 from azurectl.endpoint import Endpoint
