+++ test/unit/reserved_ip_test.py
@@ -7,7 +7,7 @@ from mock import call
 
 from test_helper import *
 
-from azurectl.azure_account import AzureAccount
+from azurectl.account.service import AzureAccount
 from azurectl.config import Config
 from azurectl.azurectl_exceptions import *
 from azurectl.reserved_ip import ReservedIp
